fixed https://bugzilla.novell.com/show_bug.cgi?id=80572
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@4519 db2431b9-df06-0410-971c-a38ea9a75485

+++ simias/src/Sync/Conflict.cs
@@ -362,8 +362,6 @@ public class Conflict
 		if (fn != null)
 		{
 			DirNode parent = fn.GetParent(collection);
-			if (SyncFile.DoesNodeExist(collection, parent, newNodeName))
-				throw new ExistsException(newNodeName);
 			if (newNodeName == node.Name)
 			{
 				// We are resolving to the same name.
@@ -377,6 +375,8 @@ public class Conflict
 			}
 			else
 			{
+				if (SyncFile.DoesNodeExist(collection, parent, newNodeName))
+					throw new ExistsException(newNodeName);
 				//TODO: what if move succeeds but node rename or commit fails?
 				File.Move(FileNameConflictPath, Path.Combine(Path.GetDirectoryName(NonconflictedPath), newNodeName));
 				string relativePath = fn.GetRelativePath();
